Moving utils, some const fn, Kyber768 parameters
